developer-wallet-tutorial/requirements.txt
circle-developer-controlled-wallets==6.1.0 